requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
bcrypt==4.1.2
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distlib==0.3.9
dnspython==2.7.0
email_validator==2.1.2
fastapi==0.115.4
fastapi-cli==0.0.5
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
identify==2.6.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
makefun==1.15.6
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.9.1
orjson==3.10.10
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pwdlib==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.9.3
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
starlette==0.41.2
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.32.0
virtualenv==20.27.1
watchfiles==0.24.0
websockets==13.1
geopy==2.4.1
fastapi-cache2[redis]
async_factory_boy==1.0.1
pytest-asyncio==0.24.0
